Moved package json helpers to utils files
