+++ backend/src/services/MetricsService.ts
@@ -1,0 +1,11 @@
+/** 
+ * @file MetricsService.ts
+ * Service contains the business logic for handling metrics data & calculation.
+ * 
+ */
+
+export class MetricsService {
+
+    
+
+}
