+++ backend/src/services/PackageUploadService.ts
@@ -4,13 +4,17 @@
  */
 
 
-import AdmZip from 'adm-zip';
+
 import axios from 'axios';
 
 import { ApiError } from "../utils/errors/ApiError.js";
 import { AppDataSource } from "../data-source.js";
 import { PackageMetadata } from "../entities/PackageMetadata.js";
 import { PackageData } from "../entities/PackageData.js";
+
+import { getPackageJsonFromContentBuffer } from '../utils/packageDataHelpers.js';
+import { extractNameAndVersionFromPackageJson } from '../utils/packageDataHelpers.js';
+
 import octokit from '../utils/octokit.js';
 
 /**
@@ -20,6 +24,7 @@ import octokit from '../utils/octokit.js';
  */
 export class PackageUploadService {
 
+    
     /**
      * @function uploadContentType
      * Uploads a package that contains Base64 encoded zip file to the database. 
@@ -32,15 +37,14 @@ export class PackageUploadService {
     static async uploadContentType(Content: string, JSProgram: string, debloat: boolean) {
         console.log('[PackageService] Uploading Content type package to the database.');
 
+        // Extract the package.json from the zip file
+        const packageJson = await getPackageJsonFromContentBuffer(Content);
+
         // Extract the package name and version from the zip file
-        const extracted = await this.extractNameAndVersionFromZip(Content);
-        if (!extracted) {
-            throw new ApiError('Failed to extract name and version from zip content', 400);
-        }
-        const { Name, Version } = extracted;
+        const { Name, Version } = extractNameAndVersionFromPackageJson(packageJson);
 
         // Get PackageMetadata repository, create metadata and save
-        const packageMetadataRepository = await AppDataSource.getRepository(PackageMetadata);
+        const packageMetadataRepository = AppDataSource.getRepository(PackageMetadata);
 
         // Check if the package already exists
         const existingMetadata = await packageMetadataRepository.findOne({ 
@@ -58,7 +62,7 @@ export class PackageUploadService {
         await packageMetadataRepository.save(metadata);
 
         // Get PackageData repository, create data and save
-        const packageDataRepository = await AppDataSource.getRepository(PackageData);
+        const packageDataRepository = AppDataSource.getRepository(PackageData);
         const data = packageDataRepository.create({
             // 1:1 relationship between metadata and data
             packageMetadata: metadata,
@@ -103,11 +107,11 @@ export class PackageUploadService {
             const response = await axios.get(normalizedURL, { responseType: 'arraybuffer' });
             const zipBuffer = Buffer.from(response.data, 'binary');
             const base64Zip = zipBuffer.toString('base64');
-            const extracted = await this.extractNameAndVersionFromZip(base64Zip);
-            if (!extracted) {
-                throw new Error('Failed to extract name and version from zip content');
-            }
-            const { Name, Version } = extracted;
+
+            const packageJson = await getPackageJsonFromContentBuffer(base64Zip);
+
+            const { Name, Version } = extractNameAndVersionFromPackageJson(packageJson);
+        
             const packageMetadataRepository = await AppDataSource.getRepository(PackageMetadata);
             const metadata = packageMetadataRepository.create({
                 name: Name,
@@ -142,52 +146,7 @@ export class PackageUploadService {
         }
     }
 
-    /**
-     * @function extractNameAndVersionFromZip
-     * Extracts the name and version of the package from the package.json file in the zip content.
-     * 
-     * @param Content Base64 encoded zip file
-     * @returns Object containing the name and version of the package
-     */
-    static async extractNameAndVersionFromZip(Content: string) {
-
-        try {
-
-            // Decode the base64 encoded zip file to binary buffer
-            const zipBuffer = Buffer.from(Content, 'base64');
-
-            // Load buffer as zip file and extract package.json
-            const zip = new AdmZip(zipBuffer);
-            const zipEntries = zip.getEntries();
-            const targetEntry = zipEntries.find(entry => entry.entryName.endsWith('package.json'));
-            
-            if (!targetEntry) {
-                throw new ApiError('Package.json not found.', 400);
-            }
-
-            // Parse package.json file
-            const fileData = targetEntry.getData();
-            const packageJson = JSON.parse(fileData.toString('utf8'));
-            
-            // Extract name and version if available
-            const Name = packageJson.name;
-            const Version = packageJson.version;
-
-            if (!Name || !Version) {
-                throw new ApiError('Name or version not found in package.json.', 400);
-            }
-
-            return { Name, Version };
-
-        } catch (error) {
-            if (error instanceof ApiError) {
-                throw error;
-            } 
-            console.error('[PackageUploadService] An error occurred while extracting the name and version from the zip content.', error);
-            throw new ApiError("Failed to extract name and version from zip content.", 400);
-        }
-    };
-
+    
     /**
      * @function normalizePackageURL
      * Converts npm link to GitHub link if applicable.
@@ -272,8 +231,6 @@ export class PackageUploadService {
             throw new ApiError('Failed to get default branch.', 400);
         }
     } // end getDefaultBranch
-
-     
 
 };
